PoC1のSlack通知を共通化し、PoC2にも適用
